# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "shotgunapp\n"
  version: "1.0.0"
  title: "Swagger Shotgun"
  termsOfService: "http://swagger.io/terms/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
# host: "localhost:8080"
# basePath: "/api"
tags:
- name: "shotgun"
  description: "Everything about your Shotguns"
- name: "event"
  description: "Access to the event shotgun"
- name: "user"
  description: "Operations about user"
- name: "game"
  description: "Everything about your games"
# schemes:
# - "http"
# - "https"
paths:
  /shotgun:
    post:
      tags:
      - "shotgun"
      summary: "Add a new shotgun"
      operationId: "createshotgun"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "add the shotgun to the list"
        required: true
        schema:
          $ref: "#/definitions/Shotgun"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "shotgun.controller"
  /shotgun/{cle}:
    get:
      tags:
      - "shotgun"
      summary: "Find shotgun by cle"
      description: "Returns a single shotgun"
      operationId: "findOneshotgun"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "id of shotgun"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Shotgun"
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Shotgun not found"
      x-swagger-router-controller: "shotgun.controller"
    put:
      tags:
      - "shotgun"
      summary: "Update an existing shotgun"
      operationId: "updateshotgun"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "id of shotgun"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "update the shotgun to the list"
        required: true
        schema:
          $ref: "#/definitions/Shotgun"
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Shotgun not found"
        "405":
          description: "Validation exception"
      x-swagger-router-controller: "shotgun.controller"
    delete:
      tags:
      - "shotgun"
      summary: "Deletes a shotgun"
      operationId: "deleteshotgun"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "id shotgun to delete"
        required: true
        type: "integer"
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Shotgun not found"
      x-swagger-router-controller: "shotgun.controller"
  /shotgun/shotlist/{username}:
    get:
      tags:
      - "shotgun"
      summary: "Find shotgun by username"
      description: "Returns list of single shotgun"
      operationId: "findListshotgun"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "username of shotgun"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: array
            items:
             $ref: "#/definitions/Shotgun"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "Shotgun List not found"
      x-swagger-router-controller: "shotgun.controller"
  /shotgun/shotlist/nom/{nom_shotgun}:
    get:
      tags:
      - "shotgun"
      summary: "Find shotgun by nom_shotgun"
      description: "Returns list of single shotgun"
      operationId: "findshotgun"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "nom_shotgun"
        in: "path"
        description: "name of shotgun"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: array
            items:
             $ref: "#/definitions/Shotgun"
        "400":
          description: "Invalid name supplied"
        "404":
          description: "Shotgun not found"
      x-swagger-router-controller: "shotgun.controller"
  /event:
    post:
      tags:
      - "event"
      summary: "Do a shotgun"
      operationId: "createevent"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "add your own shotgun of an event to the list"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Event"
  /event/{cle}:
    get:
      tags:
      - "event"
      summary: "Find all the shotgun by cle"
      description: "Returns the time of each shotgun with the same id"
      operationId: "findOneevent"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "id of shotgun"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid cle supplied"
        "404":
          description: "Shotgun not found"
      x-swagger-router-controller: "Event"
  
    put:
      tags:
      - "event"
      summary: "add something to an existing event"
      operationId: "updateevent"
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: cle
        in: path
        description: id of event
        required: true
        type: integer
      - in: body
        name: body
        description: update the event to the list
        required: true
        schema:
          $ref: '#/definitions/Event'
      responses:
        400:
          description: Invalid id supplied
        404:
          description: Shotgun not found
        405:
          description: Validation exception
    delete:
      tags:
      - "event"
      summary: "Delete shotgun of an event"
      description: "This can only be done by the logged in user."
      operationId: "deleteevent"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "The id that needs to be deleted"
        required: true
        type: integer
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/user/{username}:
    get:
      tags:
      - "event"
      summary: "Find all the event by username"
      description: "Returns each event from the same user"
      operationId: "findOneeventByUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "username of user"
        required: true
        type: "integer"
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
        "400":
          description: "Invalid cle supplied"
        "404":
          description: "Shotgun not found"
      x-swagger-router-controller: "Event"
  /event/user/:{username}/cle/:{cle}:
    get:
      tags:
      - "event"
      summary: "Find an event by cle and username"
      description: "Returns a specific event"
      operationId: "finduniqueevent"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "cle of shotgun"
        required: true
        type: "integer"
      - name: "username"
        in: "path"
        description: "username of shotgun"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid cle or username supplied"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  
    put:
      tags:
      - "event"
      summary: "add something to an existing event"
      operationId: "updateuniqueevent"
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: cle
        in: path
        description: id of event
        required: true
        type: integer
      - name: username
        in: path
        description: username of event
        required: true
        type: string
      - in: body
        name: body
        description: update the event to the list
        required: true
        schema:
          $ref: '#/definitions/Event'
      responses:
        400:
          description: Invalid cle or username supplied
        404:
          description: Event not found
        405:
          description: Validation exception
    delete:
      tags:
      - "event"
      summary: "Delete shotgun of an event"
      description: "This can only be done by the logged in user."
      operationId: "deleteuniqueevent"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "cle"
        in: "path"
        description: "The id that needs to be deleted"
        required: true
        type: integer
      - name: "username"
        in: "path"
        description: "The id and username that needs to be deleted"
        required: true
        type: string
      responses:
        "400":
          description: "Invalid id or username supplied"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createuser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "User"
  /user/{username}:
    get:
      tags:
      - "user"
      summary: "Get user by user name"
      operationId: "findOneuser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateuser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "400":
          description: "Invalid user supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteuser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
  /game:
    post:
      tags:
      - "game"
      summary: "Add a new game"
      operationId: "creategame"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Game object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Game"
      responses:
        "200":
          description: "succesfull operation"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Game"
  /game/{id_game}:
    get:
      tags:
      - "game"
      summary: "Get game by id"
      operationId: "findOnegame"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_game"
        in: "path"
        description: "The game that needs to be fetched."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Game"
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Game not found"
      x-swagger-router-controller: "Game"
    put:
      tags:
      - "game"
      summary: "Updated game"
      description: "This can only be done by the logged in user."
      operationId: "updategame"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_game"
        in: "path"
        description: "game that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated date_go object"
        required: true
        schema:
          $ref: "#/definitions/Game"
      responses:
        "400":
          description: "Invalid game_id supplied"
        "404":
          description: "Game not found"
      x-swagger-router-controller: "Game"
    delete:
      tags:
      - "game"
      summary: "Delete game"
      description: "This can only be done by the logged in user."
      operationId: "deletegame"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_game"
        in: "path"
        description: "The game id that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "game not found"
      x-swagger-router-controller: "Game"
definitions:
  User:
    type: "object"
    properties:
      id_user:
        type: "integer"
      username:
        type: "string"
      password:
        type: "string"
      fullname:
        type: "string"
      photo:
        type: "string"
      bio:
        type: "string"
      age:
        type: 'integer'
      telephone:
        type: 'integer'
      genre:
        type: 'string'
      lieu:
        type: 'string'
    xml:
      name: "User"
    example:
      id_user: 123012
      username: "username"
      password: "1234"
      fullname: "fullname"
      photo: "photo"
      bio: "bio"    
      age: "18"
      telephone: "0695954346"
      genre: "M"
      lieu: "albi"
  Shotgun:
    type: "object"
    required:
    - "nom_shotgun"
    properties:
      cle:
        type: "integer"
      nom_shotgun:
        type: "string"
      date_shotgun:
        type: "string"
        format: "date"
      nb_place:
        type: "integer"
      photo_shotgun:
        type: "string"   
      email:
        type: "boolean"
      age:
        type: "boolean"
      telephone:
        type: "boolean"
      genre:
        type: "boolean"
      custom:
        type: "boolean"
      custom_text:
        type: "string"
    xml:
      name: "Shotgun"
    example:
      cle: 154690
      nom_shotgun: "weski"
      date_shotgun: "20/10/2022"
      nb_place: 2
      photo_shotgun: "photo"
      email: "false"
      age: "false"
      telephone: "false"
      genre: "false"
      custom: "false"
      custom_text: "promo"
  Event:
    type: "object"
    properties:
      id_event:
        type: "integer"
      id_game:
        type: "integer"
      id_user:
        type: "integer"
      cle:
        type: "integer"
      username:
        type: "string"
      admin:
        type: "boolean"
      age:
        type: 'integer'
      telephone:
        type: 'integer'
      genre:
        type: 'string'
      custom:
        type: 'string'
      temps_shot:
        type: "integer"
      visible:
        type: "boolean"
    xml:
      name: "Event"
    example:
      id_event: 100000
      id_game: 100000
      id_user: 100000
      cle: 100000
      username: "username"
      admin: "true"
      age: 18
      telephone: "0695954346"
      genre: "M"
      custom: "custom"
      temps_shot: 6
      visible: "true"
      
  Game:
    type: "object"
    properties:
      id_game:
        type: "integer"
      username:
        type: "string"
      admin:
        type: "boolean"
      temps_jeu:
        type: "integer"
      date_go:
        type: "string"
        format: "date"
    xml:
      name: "Game"
    example:
      id_game: 100650
      username: "username"
      admin: "true"
      temps_jeu: 6
      date_go: "date_go"
      
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /Tom_s_Industry/ShotgunApp_Api/1.0.0
schemes:
 - https
 - http